Now calling the application's exception handler directly, instead of relying on a more indirect or magical approach of exception handling by throwing the exception up the call-stack.
This fixes an issue where exceptions were only being handled by the PHP
runtime's registered exception handler as a "last resort", which
unfortunately means that the application's execution halts immediately
afterwards.

This not only fixes that issue, though, as it also improves code-path
detection and readability, as well as testability, as the exception's
path of handling is now more explicit, while still behaving in the same
manner as before.